+++ rust/dbn/src/enums.rs
@@ -641,6 +641,11 @@ pub enum StatType {
     /// The volume-weighted average price (VWAP) for a fixing period. `price` will be
     /// set.
     FixingPrice = 10,
+    /// The last trade price during a trading session. `price` will be set.
+    ClosePrice = 11,
+    /// The change in price from the close price of the previous trading session to the
+    /// most recent trading session. `price` will be set.
+    NetChange = 12,
 }
 
 /// The type of [`StatMsg`](crate::record::StatMsg) update.